+++ protocol-designer/src/components/StepEditForm/formFields.js
@@ -12,13 +12,14 @@ import {
 import i18n from '../../localization'
 import {selectors as pipetteSelectors} from '../../pipettes'
 import {selectors as labwareIngredSelectors} from '../../labware-ingred/reducers'
-import {actions} from '../../steplist'
 import {hydrateField} from '../../steplist/fieldLevel'
 import type {StepFieldName} from '../../steplist/fieldLevel'
-import {DISPOSAL_PERCENTAGE} from '../../steplist/formLevel/warnings'
-import {SOURCE_WELL_DISPOSAL_DESTINATION} from '../../steplist/formLevel/stepFormToArgs/transferLikeFormToArgs'
+import {
+  SOURCE_WELL_BLOWOUT_DESTINATION,
+  DEST_WELL_BLOWOUT_DESTINATION,
+} from '../../step-generation/utils'
 import type {ChangeTipOptions} from '../../step-generation/types'
-import type {BaseState, ThunkDispatch} from '../../types'
+import type {BaseState} from '../../types'
 import type {StepType} from '../../form-types'
 import styles from './StepEditForm.css'
 import StepField from './StepFormField'
@@ -96,40 +97,14 @@ export const StepRadioGroup = (props: StepRadioGroupProps) => {
   )
 }
 
-type DispenseDelayFieldsProps = {
-  focusHandlers: FocusHandlers,
-  label?: string,
-  disabled?: boolean,
-  tooltipComponent?: React.Node,
-}
-export function DispenseDelayFields (props: DispenseDelayFieldsProps) {
-  const {label = 'Delay', focusHandlers, tooltipComponent, disabled} = props
-  return (
-    <StepCheckboxRow
-      disabled={disabled}
-      tooltipComponent={tooltipComponent}
-      name="dispense_delay_checkbox"
-      label={label}>
-      <StepInputField {...focusHandlers} disabled={disabled} name="dispense_delayMinutes" units='m' />
-      <StepInputField {...focusHandlers} disabled={disabled} name="dispense_delaySeconds" units='s' />
-    </StepCheckboxRow>
-  )
-}
-
 type PipetteFieldOP = {name: StepFieldName, stepType?: StepType} & FocusHandlers
 type PipetteFieldSP = {pipetteOptions: Options, getHydratedPipette: (string) => any} // TODO: real hydrated pipette type
-type PipetteFieldDP = {updateDisposalVolume: (?mixed) => void}
-type PipetteFieldProps = PipetteFieldOP & PipetteFieldSP & PipetteFieldDP
+type PipetteFieldProps = PipetteFieldOP & PipetteFieldSP
 const PipetteFieldSTP = (state: BaseState, ownProps: PipetteFieldOP): PipetteFieldSP => ({
   pipetteOptions: pipetteSelectors.getEquippedPipetteOptions(state),
   getHydratedPipette: (value) => hydrateField(state, ownProps.name, value),
 })
-const PipetteFieldDTP = (dispatch: ThunkDispatch<*>): PipetteFieldDP => ({
-  updateDisposalVolume: (disposalVolume: ?mixed) => {
-    dispatch(actions.changeFormInput({update: {aspirate_disposalVol_volume: disposalVolume}}))
-  },
-})
-export const PipetteField = connect(PipetteFieldSTP, PipetteFieldDTP)((props: PipetteFieldProps) => (
+export const PipetteField = connect(PipetteFieldSTP)((props: PipetteFieldProps) => (
   <StepField
     name={props.name}
     focusedField={props.focusedField}
@@ -143,23 +118,35 @@ export const PipetteField = connect(PipetteFieldSTP, PipetteFieldDTP)((props: Pi
           onFocus={() => { props.onFieldFocus(props.name) }}
           onChange={(e: SyntheticEvent<HTMLSelectElement>) => {
             updateValue(e.currentTarget.value)
-            if (props.stepType === 'distribute') {
-              const hydratedPipette = props.getHydratedPipette(e.currentTarget.value)
-              if (hydratedPipette) {
-                props.updateDisposalVolume(hydratedPipette.maxVolume * DISPOSAL_PERCENTAGE)
-              }
-            }
           }} />
       </FormGroup>
     )} />
 ))
 
-type DisposalDestinationDropdownOP = {name: StepFieldName, className?: string} & FocusHandlers
-type DisposalDestinationDropdownSP = {options: Options}
-const DisposalDestinationDropdownSTP = (state: BaseState): DisposalDestinationDropdownSP => ({
-  options: labwareIngredSelectors.disposalLabwareOptions(state),
-})
-export const DisposalDestinationDropdown = connect(DisposalDestinationDropdownSTP)((props: DisposalDestinationDropdownOP & DisposalDestinationDropdownSP) => {
+type BlowoutLocationDropdownOP = {
+  name: StepFieldName,
+  className?: string,
+  includeSourceWell?: ?boolean,
+  includeDestWell?: ?boolean,
+} & FocusHandlers
+type BlowoutLocationDropdownSP = {options: Options}
+const BlowoutLocationDropdownSTP = (state: BaseState, ownProps: BlowoutLocationDropdownOP): BlowoutLocationDropdownSP => {
+  let options = labwareIngredSelectors.disposalLabwareOptions(state)
+  if (ownProps.includeDestWell) {
+    options = [
+      ...options,
+      {name: 'Destination Well', value: DEST_WELL_BLOWOUT_DESTINATION},
+    ]
+  }
+  if (ownProps.includeSourceWell) {
+    options = [
+      ...options,
+      {name: 'Source Well', value: SOURCE_WELL_BLOWOUT_DESTINATION},
+    ]
+  }
+  return {options}
+}
+export const BlowoutLocationDropdown = connect(BlowoutLocationDropdownSTP)((props: BlowoutLocationDropdownOP & BlowoutLocationDropdownSP) => {
   const {options, name, className, focusedField, dirtyFields, onFieldBlur, onFieldFocus} = props
   return (
     <StepField
@@ -169,7 +156,7 @@ export const DisposalDestinationDropdown = connect(DisposalDestinationDropdownST
       render={({value, updateValue}) => (
         <DropdownField
           className={className}
-          options={[...options, {name: 'Source Well', value: SOURCE_WELL_DISPOSAL_DESTINATION}]}
+          options={options}
           onBlur={() => { onFieldBlur(name) }}
           onFocus={() => { onFieldFocus(name) }}
           value={value ? String(value) : null}