+++ protocol-designer/src/localization/en/index.js
@@ -3,6 +3,7 @@
 import alert from './alert.json'
 import button from './button.json'
 import card from './card.json'
+import context_menu from './context_menu.json'
 import deck from './deck.json'
 import form from './form.json'
 import modal from './modal.json'
@@ -15,6 +16,7 @@ export default {
     alert,
     button,
     card,
+    context_menu,
     deck,
     form,
     modal,