+++ protocol-designer/src/step-generation/commandCreators/compound/consolidate.js
@@ -1,11 +1,11 @@
 // @flow
 import chunk from 'lodash/chunk'
 import flatMap from 'lodash/flatMap'
-import {FIXED_TRASH_ID} from '../../../constants'
 import * as errorCreators from '../../errorCreators'
 import {getPipetteWithTipMaxVol} from '../../robotStateSelectors'
 import type {ConsolidateFormData, RobotState, CommandCreator, CompoundCommandCreator} from '../../types'
-import {aspirate, dispense, blowout, replaceTip, touchTip} from '../atomic'
+import {blowoutUtil} from '../../utils'
+import {aspirate, dispense, replaceTip, touchTip} from '../atomic'
 import {mixUtil} from './mix'
 
 const consolidate = (data: ConsolidateFormData): CompoundCommandCreator => (prevRobotState: RobotState) => {
@@ -40,13 +40,8 @@ const consolidate = (data: ConsolidateFormData): CompoundCommandCreator => (prev
     dispenseOffsetFromBottomMm,
   } = data
 
-  // TODO error on negative data.disposalVolume?
-  const disposalVolume = (data.disposalVolume && data.disposalVolume > 0)
-    ? data.disposalVolume
-    : 0
-
   const maxWellsPerChunk = Math.floor(
-    (getPipetteWithTipMaxVol(data.pipette, prevRobotState) - disposalVolume) / data.volume
+    getPipetteWithTipMaxVol(data.pipette, prevRobotState) / data.volume
   )
 
   const commandCreators = flatMap(
@@ -54,8 +49,6 @@ const consolidate = (data: ConsolidateFormData): CompoundCommandCreator => (prev
     (sourceWellChunk: Array<string>, chunkIndex: number): Array<CommandCreator> => {
       // Aspirate commands for all source wells in the chunk
       const aspirateCommands = flatMap(sourceWellChunk, (sourceWell: string, wellIndex: number): Array<CommandCreator> => {
-        const isFirstWellInChunk = wellIndex === 0
-
         const touchTipAfterAspirateCommand = data.touchTipAfterAspirate
           ? [touchTip({
             pipette: data.pipette,
@@ -68,7 +61,7 @@ const consolidate = (data: ConsolidateFormData): CompoundCommandCreator => (prev
         return [
           aspirate({
             pipette: data.pipette,
-            volume: data.volume + (isFirstWellInChunk ? disposalVolume : 0),
+            volume: data.volume,
             labware: data.sourceLabware,
             well: sourceWell,
             offsetFromBottomMm: aspirateOffsetFromBottomMm,
@@ -93,16 +86,6 @@ const consolidate = (data: ConsolidateFormData): CompoundCommandCreator => (prev
           well: data.destWell,
           offsetFromBottomMm: data.touchTipAfterDispenseOffsetMmFromBottom,
         })]
-        : []
-
-      const trashTheDisposalVol = disposalVolume
-        ? [
-          blowout({
-            pipette: data.pipette,
-            labware: FIXED_TRASH_ID,
-            well: 'A1',
-          }),
-        ]
         : []
 
       const mixBeforeCommands = (data.mixFirstAspirate)
@@ -142,15 +125,14 @@ const consolidate = (data: ConsolidateFormData): CompoundCommandCreator => (prev
         )
         : []
 
-      const blowoutCommand = (data.blowout)
-        ? [
-          blowout({
-            pipette: data.pipette,
-            labware: data.blowout, // TODO Ian 2018-05-04 more explicit test for non-trash blowout destination
-            well: 'A1', // TODO LATER: should user be able to specify the blowout well?
-          }),
-        ]
-        : []
+      const blowoutCommand = blowoutUtil(
+        data.pipette,
+        data.sourceLabware,
+        sourceWellChunk[0],
+        data.destLabware,
+        data.destWell,
+        data.blowoutLocation,
+      )
 
       return [
         ...tipCommands,
@@ -165,7 +147,6 @@ const consolidate = (data: ConsolidateFormData): CompoundCommandCreator => (prev
           offsetFromBottomMm: dispenseOffsetFromBottomMm,
         }),
         ...touchTipAfterDispenseCommands,
-        ...trashTheDisposalVol,
         ...mixAfterCommands,
         ...blowoutCommand,
       ]