+++ protocol-designer/src/step-generation/types.js
@@ -38,8 +38,6 @@ export type TransferLikeFormDataFields = {
   touchTipAfterAspirateOffsetMmFromBottom?: ?number,
   /** changeTip is interpreted differently by different Step types */
   changeTip: ChangeTipOptions,
-  /** Disposal volume is added to the volume of the first aspirate of each asp-asp-disp cycle */
-  disposalVolume: ?number,
   /** offset from bottom of well in mm */
   aspirateOffsetFromBottomMm?: ?number,
 
@@ -48,8 +46,6 @@ export type TransferLikeFormDataFields = {
   touchTipAfterDispense: boolean,
   /** Optional offset for touch tip after dispense (if null, use PD default) */
   touchTipAfterDispenseOffsetMmFromBottom?: ?number,
-  /** Number of seconds to delay at the very end of the step (TODO: or after each dispense ?) */
-  delayAfterDispense: ?number,
   /** offset from bottom of well in mm */
   dispenseOffsetFromBottomMm?: ?number,
 }
@@ -60,8 +56,9 @@ export type ConsolidateFormData = {
   sourceWells: Array<string>,
   destWell: string,
 
-  /** If given, blow out in the specified labware after dispense at the end of each asp-asp-dispense cycle */
-  blowout: ?string, // TODO LATER LabwareId export type here instead of string?
+  /** If given, blow out in the specified destination after dispense at the end of each asp-asp-dispense cycle */
+  blowoutLocation: ?string,
+
   /** Mix in first well in chunk */
   mixFirstAspirate: ?MixArgs,
   /** Mix in destination well after dispense */
@@ -74,8 +71,9 @@ export type TransferFormData = {
   sourceWells: Array<string>,
   destWells: Array<string>,
 
-  /** If given, blow out in the specified labware after dispense at the end of each asp-asp-dispense cycle */
-  blowout: ?string, // TODO LATER LabwareId export type here instead of string?
+  /** If given, blow out in the specified destination after dispense at the end of each asp-dispense cycle */
+  blowoutLocation: ?string,
+
   /** Mix in first well in chunk */
   mixBeforeAspirate: ?MixArgs,
   /** Mix in destination well after dispense */
@@ -88,6 +86,8 @@ export type DistributeFormData = {
   sourceWell: string,
   destWells: Array<string>,
 
+  /** Disposal volume is added to the volume of the first aspirate of each asp-asp-disp cycle */
+  disposalVolume: ?number,
   /** Disposal labware and well for final blowout destination of disposalVolume contents (e.g. trash, source well, etc.) */
   disposalLabware: ?string,
   disposalWell: ?string,
@@ -109,12 +109,12 @@ export type MixFormData = {
   /** Touch tip after mixing */
   touchTip: boolean,
   touchTipMmFromBottom?: ?number,
-  /** Delay in seconds */
-  delay: ?number,
   /** change tip: see comments in step-generation/mix.js */
   changeTip: ChangeTipOptions,
-  /** If given, blow out in the specified labware after mixing each well */
-  blowout?: string,
+
+  /** If given, blow out in the specified destination after mixing each well */
+  blowoutLocation: ?string,
+
   /** offset from bottom of well in mm */
   aspirateOffsetFromBottomMm?: ?number,
   dispenseOffsetFromBottomMm?: ?number,