+++ protocol-designer/src/step-generation/utils.js
@@ -12,10 +12,15 @@ import type {
   RobotState,
   SourceAndDest,
   Timeline,
+  PipetteLabwareFields,
 } from './types'
+import blowout from './blowout'
 
 import {AIR} from '@opentrons/components'
 export {AIR}
+
+export const SOURCE_WELL_BLOWOUT_DESTINATION: 'source_well' = 'source_well'
+export const DEST_WELL_BLOWOUT_DESTINATION: 'dest_well' = 'dest_well'
 
 export function repeatArray<T> (array: Array<T>, repeats: number): Array<T> {
   return flatMap(range(repeats), (i: number): Array<T> => array)
@@ -205,4 +210,27 @@ export function totalVolume (location: LocationLiquidState): number {
       ? acc + (location[ingredId].volume || 0)
       : acc
   }, 0)
+}
+
+export const blowoutUtil = (
+  pipette: $PropertyType<PipetteLabwareFields, 'pipette'>,
+  sourceLabware: $PropertyType<PipetteLabwareFields, 'labware'>,
+  sourceWell: $PropertyType<PipetteLabwareFields, 'well'>,
+  destLabware: $PropertyType<PipetteLabwareFields, 'labware'>,
+  destWell: $PropertyType<PipetteLabwareFields, 'well'>,
+  blowoutLocation: ?string,
+): Array<CommandCreator> => {
+  if (!blowoutLocation) return []
+  let labware = blowoutLocation
+  let well = 'A1'
+
+  // TODO Ian 2018-05-04 more explicit test for non-trash blowout destination
+  if (blowoutLocation === SOURCE_WELL_BLOWOUT_DESTINATION) {
+    labware = sourceLabware
+    well = sourceWell
+  } else if (blowoutLocation === DEST_WELL_BLOWOUT_DESTINATION) {
+    labware = destLabware
+    well = destWell
+  }
+  return [blowout({pipette: pipette, labware, well})]
 }